requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
python-multipart>=0.0.12

# Database — async SQLAlchemy + pgvector
sqlmodel>=0.0.21
sqlalchemy[asyncio]>=2.0.35
asyncpg>=0.29.0
pgvector>=0.3.2

# Migrations (uses sync psycopg2 for Alembic connection)
alembic>=1.13.3
psycopg2-binary>=2.9.9

# Config
pydantic-settings>=2.4.0

# Auth
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# PDF parsing
pdfplumber>=0.11.4

# Tokenization
tiktoken>=0.7.0

# AI APIs
openai>=1.51.0
anthropic>=0.34.2

# Testing
pytest>=8.3.3
pytest-asyncio>=0.24.0
httpx>=0.27.2
